feat/add-document-chunks

+++ app/schemas/dashboard.py
@@ -6,4 +6,4 @@ class DashboardSummary(BaseModel):
     processed_documents: int
     failed_documents: int
     storage_bytes: int
-    detected_languages: dict[str, int]
+    detected_languages: dict[str, float]
